Change electron-app.ts to app-electron.ts and move menu.ts to main folder.
